fix: pull fixLocations out into its own typescript file
This also makes parser.js trivial, so I converted it to a typescript file as
well.
