--- src/util/type.ts
@@ -1,4 +1,0 @@
-// tslint:disable-next-line:no-any
-export default function type(node: any): string {
-  return node.constructor.name;
-}
